refactor(web): move session detail UI into features/sessions
Give ModernSession its own feature owned by the /sessions pages, with a
real index.ts front door for annotation-queues consumers. Pages import
SessionPage/SessionEventsPage directly (same pattern as TracePage).
