+++ web/src/features/sessions/ModernSessionSidebar.tsx
@@ -12,17 +12,17 @@ import {
 import { type FilterState } from "@langfuse/shared";
 
 import { renderFilterIcon } from "@/src/components/ItemBadge";
-import { SessionVirtualizedRow } from "@/src/features/annotation-queues/components/session/SessionVirtualizedRow";
-import { type EventSessionTrace } from "@/src/features/annotation-queues/components/session/sessionDetailPageTypes";
+import { SessionVirtualizedRow } from "@/src/features/sessions/SessionVirtualizedRow";
+import { type EventSessionTrace } from "@/src/features/sessions/sessionDetailPageTypes";
 import {
   ModernSessionViewDropdownMenu,
   type ModernSessionViewDropdownMenuControls,
-} from "@/src/features/annotation-queues/components/session/ModernSessionViewDropdownMenu";
-import { SESSION_DETAIL_VIEW_TRIGGER_ID } from "@/src/features/annotation-queues/components/session/session-detail-presets";
+} from "@/src/features/sessions/ModernSessionViewDropdownMenu";
+import { SESSION_DETAIL_VIEW_TRIGGER_ID } from "@/src/features/sessions/session-detail-presets";
 import {
   formatIdleGap,
   IDLE_GAP_THRESHOLD_SECONDS,
-} from "@/src/features/annotation-queues/components/session/sessionIdleGap";
+} from "@/src/features/sessions/sessionIdleGap";
 import { Input } from "@/src/components/ui/input";
 import { Button } from "@/src/components/ui/button";
 import {
