+++ web/src/features/sessions/SessionMetadataJsonPathControl.tsx
@@ -16,8 +16,8 @@ import {
   metadataJsonPathsStorageKey,
   parseStoredMetadataJsonPaths,
   type SessionMetadataJsonPathState,
-} from "@/src/features/annotation-queues/components/session/sessionMetadataJsonPath";
-import { getVisibleSessionObservations } from "@/src/features/annotation-queues/components/session/sessionVisibleObservations";
+} from "@/src/features/sessions/sessionMetadataJsonPath";
+import { getVisibleSessionObservations } from "@/src/features/sessions/sessionVisibleObservations";
 import useLocalStorage from "@/src/components/useLocalStorage";
 import { api, type RouterOutputs } from "@/src/utils/api";
