+++ web/src/features/sessions/session-detail-presets.ts
