+++ web/src/features/sessions/sessionHeaderVisibility.ts
