+++ web/src/features/sessions/stableVirtualRowMeasurementState.clienttest.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it, vi } from "vitest";
-import { createStableVirtualRowMeasurementState } from "@/src/features/annotation-queues/components/session/stableVirtualRowMeasurementState";
+import { createStableVirtualRowMeasurementState } from "@/src/features/sessions/stableVirtualRowMeasurementState";
 
 describe("createStableVirtualRowMeasurementState", () => {
   it("clears stale pending height when a direct commit lands", () => {
