+++ web/src/pages/project/[projectId]/sessions/[sessionId].tsx
@@ -1,7 +1,7 @@
 import {
   SessionEventsPage,
   SessionPage,
-} from "@/src/features/annotation-queues/components/session";
+} from "@/src/features/sessions/SessionPages";
 import { useReadPath } from "@/src/features/events/hooks/useReadPath";
 import {
   RouteParamsPendingFallback,
